requirements.txt
openai-agents
rich
pydantic
python-dotenv